requirements.txt
aioredis==1.3.1
amqp==5.2.0
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==3.0.2
Django==5.0.6
django-compat==1.0.15
django-cors-headers==4.3.1
django-crontab==0.7.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-guardian==2.4.0
django-picklefield==3.2
django-polymorphic==3.1.0
django-rest-framework==0.1.0
django-tinymce==4.1.0
djangorestframework==3.15.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
flower==2.0.1
hiredis==3.0.0
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
packaging==24.1
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pymemcache==3.5.2
pyOpenSSL==24.2.1
pysolr==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
schedule==1.2.2
service-identity==24.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
tornado==6.4.1
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
